Refactor: To Bloc, move analysis models and screen to dedicated files
This commit refactors the analysis feature by:

- Moving the analysis data models (BaseAnalysis, DailyAnalysis, WeeklyAnalysis, MonthlyAnalysis, YearAnalysis) from `analysis_screen.dart` to a new file `lib/data/model/analysis_model.dart`.
- Moving the `AnalysisScreen` widget from `lib/presentation/screens/analysis/analysis_screen.dart` to a new file `lib/presentation/screens/analysis/analysis/analysis_screen.dart`.
- The original `analysis_screen.dart` has been deleted.

This change improves code organization and separation of concerns.
